+++ src/main/java/model/game_space/GameStatistics.java
@@ -4,7 +4,7 @@ import model.game_entities.enums.ShieldType;
 import model.game_running.ShieldHandler;
 
 public class GameStatistics {
-    int health; //might be doubles?
+    double health; //might be doubles?
     double score;
     GameTimer timer;
     GameStatisticsListener statisticsListener;
@@ -26,7 +26,7 @@ public class GameStatistics {
      *
      * @param health the value of the player's health after the update
      */
-    public void updateHealth(int health) {
+    public void updateHealth(double health) {
         statisticsListener.onHealthChanged(health);
     }
 
@@ -60,7 +60,7 @@ public class GameStatistics {
      * a listener to notify the statistics window UI of any change to the statistics
      */
     public interface GameStatisticsListener {
-        void onHealthChanged(int health);
+        void onHealthChanged(double health);
 
         void onTimerChanged(String currentTime);
